+++ AdmisionsService/Presentation/Controllers/ManagerFacultyController.cs
@@ -1,4 +1,5 @@
-﻿using AdmisionsService.Application.Dtos;
+﻿using System.Security.Claims;
+using AdmisionsService.Application.Dtos;
 using AdmisionsService.Application.ManagerFacultyService;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -14,17 +15,27 @@ public class ManagerFacultyController : ControllerBase
     {
         _mfService = mfService;
     }
-    //[Authorize(Roles = "Manager")]
+    [Authorize(Roles = "Manager,HeadManager,Admin")]
     [HttpPost]
     public async Task<IActionResult> PostManagerFaculty(ManagerFacultyDto managerFaculty)
     {
+        var id = Guid.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+        var role = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (id != managerFaculty.ManagerId && role == "Manager")
+            return Forbid();
+        
         await _mfService.CreateManagerFaculty(managerFaculty);
         return Ok();
     }
-    //[Authorize(Roles = "Manager")]
+    [Authorize(Roles = "Manager,HeadManager,Admin")]
     [HttpDelete]
     public async Task<IActionResult> DeleteManagerFaculty(ManagerFacultyDto managerFaculty)
     {
+        var id = Guid.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+        var role = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (id != managerFaculty.ManagerId && role == "Manager")
+            return Forbid();
+        
         await _mfService.DeleteManagerFaculty(managerFaculty);
         return NoContent();
     }
